+++ packages/gov-ui-kit/CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Update style of `ProposalActionItem` module component, render link to block explorer for target address
 - Rename `tokenPrice` property to `amountUsd` on `<TransactionDataListItem />` module component
 - Update minor and patch NPM dependencies
 - Update `Yarn` version to 4.6.0